Move EventPerson mailing to more general place
