+++ indico/modules/events/management/blueprint.py
@@ -16,7 +16,7 @@
 
 from __future__ import unicode_literals
 
-from indico.modules.events.management.controllers import RHDeleteEvent, RHLockEvent, RHUnlockEvent
+from indico.modules.events.management.controllers import RHDeleteEvent, RHLockEvent, RHUnlockEvent, RHEmailEventPersons
 from indico.web.flask.wrappers import IndicoBlueprint
 
 
@@ -27,3 +27,4 @@ _bp = IndicoBlueprint('event_management', __name__, template_folder='templates',
 _bp.add_url_rule('/delete', 'delete', RHDeleteEvent, methods=('GET', 'POST'))
 _bp.add_url_rule('/lock', 'lock', RHLockEvent, methods=('GET', 'POST'))
 _bp.add_url_rule('/unlock', 'unlock', RHUnlockEvent, methods=('POST',))
+_bp.add_url_rule('/manage/person-list/email', 'email_event_persons', RHEmailEventPersons, methods=('POST',))
