+++ indico/modules/events/sessions/blueprint.py
@@ -20,8 +20,7 @@ from indico.modules.events.sessions.controllers.management.protection import RHS
 from indico.modules.events.sessions.controllers.management.sessions import (RHSessionsList, RHCreateSession,
                                                                             RHModifySession, RHDeleteSessions,
                                                                             RHExportSessionsCSV, RHExportSessionsPDF,
-                                                                            RHSessionREST, RHSessionPersonList,
-                                                                            RHSessionsEmailPersons)
+                                                                            RHSessionREST, RHSessionPersonList)
 from indico.web.flask.wrappers import IndicoBlueprint
 
 
@@ -36,8 +35,6 @@ _bp.add_url_rule('/manage/sessions/sessions.pdf', 'export_pdf', RHExportSessions
 _bp.add_url_rule('/manage/sessions/<int:session_id>', 'session_rest', RHSessionREST, methods=('PATCH', 'DELETE'))
 _bp.add_url_rule('/manage/sessions/<int:session_id>/modify', 'modify_session', RHModifySession, methods=('GET', 'POST'))
 _bp.add_url_rule('/manage/sessions/person-list/', 'person_list', RHSessionPersonList, methods=('POST',))
-_bp.add_url_rule('/manage/sessions/person-list/email', 'email_session_persons', RHSessionsEmailPersons,
-                 methods=('POST',))
 
 _bp.add_url_rule('/manage/sessions/coordinator-privs/', 'coordinator_privs_rest', RHSessionCoordinatorPrivs)
 _bp.add_url_rule('/manage/sessions/coordinator-privs/<priv>', 'coordinator_privs_rest', RHSessionCoordinatorPrivs,
