+++ indico/web/assets/bundles.py
@@ -390,6 +390,7 @@ module_js = {
     'vc': rjs_bundle('modules_vc', 'js/indico/modules/vc.js'),
     'event_display': rjs_bundle('modules_event_display', 'js/indico/modules/eventdisplay.js'),
     'event_layout': rjs_bundle('modules_event_layout', 'js/indico/modules/eventlayout.js'),
+    'event_management': rjs_bundle('modules_event_management', 'js/indico/modules/eventmanagement.js'),
     'attachments': rjs_bundle('modules_attachments', 'js/indico/modules/attachments.js'),
     'surveys': rjs_bundle('modules_surveys', 'js/indico/modules/surveys.js'),
     'registration': rjs_bundle('modules_registration',
